refactor: consolidate pedestrian environment
